Usable Item Cleanup.  extra Gizmos for Weapons
Removed UsableItem to remove redundancy with IUsableItem.
Item is now the base class for all Items, Usable or not.
The class is currently Empty and inherits from Monobehaviour.

Added Gizmos for Weapon that draws a line to any Hit Colliders and draws a small red sphere on it, as well.